# backend/docs/swagger.yaml
basePath: /api
definitions:
  backend_internal_model.Category:
    properties:
      children:
        description: 由 Repository 层手动填充
        items:
          $ref: '#/definitions/backend_internal_model.Category'
        type: array
      coverUrl:
        type: string
      createdAt:
        type: string
      deletedAt:
        format: date-time
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      parentId:
        description: 支持父子分类
        type: integer
      postCount:
        type: integer
      slug:
        type: string
      updatedAt:
        type: string
    type: object
...
  backend_internal_model.Comment:
    properties:
      content:
        type: string
      createdAt:
        type: string
      deletedAt:
        format: date-time
        type: string
      id:
        type: integer
      likes:
        type: integer
      parentId:
        description: nil = 顶级评论
        type: integer
      postId:
        type: integer
      replies:
        items:
          $ref: '#/definitions/backend_internal_model.Comment'
        type: array
      updatedAt:
        type: string
      user:
        allOf:
        - $ref: '#/definitions/backend_internal_model.UserBrief'
        description: 以下字段不写入数据库，由 Repository 层手动批量查询后填充
      userId:
        type: integer
    type: object
  backend_internal_model.CreateCategoryReq:
    properties:
      coverUrl:
        maxLength: 512
        type: string
      description:
        maxLength: 500
        type: string
      name:
        maxLength: 64
        type: string
      parentId:
        type: integer
      slug:
        maxLength: 64
        type: string
    required:
    - name
    type: object
  backend_internal_model.CreateCommentReq:
    properties:
      content:
        maxLength: 2000
        minLength: 1
        type: string
      parentId:
        type: integer
      postId:
        type: integer
    required:
    - content
    - postId
    type: object
...
  backend_internal_model.LoginReq:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
...
  backend_internal_model.RegisterReq:
    properties:
      displayName:
        maxLength: 64
        type: string
      email:
        type: string
      password:
        maxLength: 64
        minLength: 6
        type: string
      username:
        maxLength: 32
        minLength: 3
        type: string
    required:
    - email
    - password
    - username
    type: object
...
  backend_internal_model.UpdateCategoryReq:
    properties:
      coverUrl:
        maxLength: 512
        type: string
      description:
        maxLength: 500
        type: string
      name:
        maxLength: 64
        type: string
      parentId:
        type: integer
      slug:
        maxLength: 64
        type: string
    type: object
  backend_internal_model.UpdateCommentReq:
    properties:
      content:
        maxLength: 2000
        minLength: 1
        type: string
    required:
    - content
    type: object
...
  backend_internal_model.UserBrief:
    properties:
      displayName:
        type: string
      id:
        type: integer
      photoURL:
        type: string
      username:
        type: string
    type: object
  backend_internal_model.UserResponse:
    properties:
      about:
        type: string
      address:
        type: string
      city:
        type: string
      country:
        type: string
      createdAt:
        type: string
      displayName:
        type: string
      email:
        type: string
      id:
        type: integer
      isPublic:
        type: boolean
      phoneNumber:
        type: string
      photoURL:
        type: string
      role:
        type: string
      state:
        type: string
      status:
        type: string
      username:
        type: string
      zipCode:
        type: string
    type: object
  backend_internal_pkg.PageResponse:
    properties:
      code:
        example: 0
        type: integer
      data:
        $ref: '#/definitions/backend_internal_pkg.PageResult'
      msg:
        example: success
        type: string
    type: object
  backend_internal_pkg.PageResult:
    properties:
      list: {}
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
  backend_internal_pkg.Response:
    properties:
      code:
        example: 0
        type: integer
      data: {}
      msg:
        example: success
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: 基于 Gin + GORM + SQLite 的个人博客后端接口文档
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: 个人博客 API
  version: "1.0"
paths:
  /admin/categories:
    post:
      consumes:
      - application/json
      parameters:
      - description: 分类信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/backend_internal_model.CreateCategoryReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/backend_internal_pkg.Response'
            - properties:
                data:
                  $ref: '#/definitions/backend_internal_model.Category'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/backend_internal_pkg.Response'
      security:
      - BearerAuth: []
      summary: 创建分类
      tags:
      - categories
  /admin/categories/{id}:
    delete:
      parameters:
      - description: 分类ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/backend_internal_pkg.Response'
      security:
      - BearerAuth: []
      summary: 删除分类
      tags:
      - categories
    put:
      consumes:
      - application/json
      parameters:
      - description: 分类ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新内容
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/backend_internal_model.UpdateCategoryReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/backend_internal_pkg.Response'
            - properties:
                data:
                  $ref: '#/definitions/backend_internal_model.Category'
              type: object
      security:
      - BearerAuth: []
      summary: 更新分类
      tags:
      - categories
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: 账号密码登录，返回 JWT Token
      parameters:
      - description: 登录信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/backend_internal_model.LoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/backend_internal_pkg.Response'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/backend_internal_pkg.Response'
      summary: 用户登录
      tags:
      - auth
  /auth/logout:
    post:
      description: 使当前 Token 失效（递增 token_version）
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/backend_internal_pkg.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/backend_internal_pkg.Response'
      security:
      - BearerAuth: []
      summary: 登出
      tags:
      - auth
  /auth/refresh:
    post:
      description: 使用当前有效 Token 换取新 Token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/backend_internal_pkg.Response'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/backend_internal_pkg.Response'
      security:
      - BearerAuth: []
      summary: 刷新 Token
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: 注册新用户，返回用户信息
      parameters:
      - description: 注册信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/backend_internal_model.RegisterReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/backend_internal_pkg.Response'
            - properties:
                data:
                  $ref: '#/definitions/backend_internal_model.UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/backend_internal_pkg.Response'
      summary: 用户注册
      tags:
      - auth
  /categories:
    get:
      description: 分页查询顶级分类，含子分类和文章数
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/backend_internal_pkg.PageResponse'
      summary: 分类列表（公开）
      tags:
      - categories
  /categories/{id}:
    get:
      parameters:
      - description: 分类ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/backend_internal_pkg.Response'
            - properties:
                data:
                  $ref: '#/definitions/backend_internal_model.Category'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/backend_internal_pkg.Response'
      summary: 分类详情（公开）
      tags:
      - categories
...
  /v1/comments:
    post:
      consumes:
      - application/json
      parameters:
      - description: 评论内容
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/backend_internal_model.CreateCommentReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/backend_internal_pkg.Response'
            - properties:
                data:
                  $ref: '#/definitions/backend_internal_model.Comment'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/backend_internal_pkg.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/backend_internal_pkg.Response'
      security:
      - BearerAuth: []
      summary: 发布评论
      tags:
      - comments
  /v1/comments/{id}:
    delete:
      parameters:
      - description: 评论ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/backend_internal_pkg.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/backend_internal_pkg.Response'
      security:
      - BearerAuth: []
      summary: 删除评论
      tags:
      - comments
    put:
      consumes:
      - application/json
      parameters:
      - description: 评论ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新内容
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/backend_internal_model.UpdateCommentReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/backend_internal_pkg.Response'
            - properties:
                data:
                  $ref: '#/definitions/backend_internal_model.Comment'
              type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/backend_internal_pkg.Response'
      security:
      - BearerAuth: []
      summary: 更新评论
      tags:
      - comments
  /v1/comments/{id}/like:
    post:
      description: 重复调用自动切换点赞状态
      parameters:
      - description: 评论ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/backend_internal_pkg.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/backend_internal_pkg.Response'
      security:
      - BearerAuth: []
      summary: 点赞/取消点赞评论
      tags:
      - comments
...
securityDefinitions:
  BearerAuth:
    description: '格式: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"